Add files via upload
This 1.0 update of this mod introduces a new item tier named rainbowite in the misc. tab of the creative menu. (in this update there is no way to get this block and these tools in plain survival without commands.)  if enough people download i will put out the whole mod mdk. PLEASE DO NOT TRY TO OPEN THE .JAR FILE I DO NOT WANT PEOPLE TO GET THE CODE OF MY MOD. That aside if you do not know how to download i will tell you the needed packages and where to get them. if you know how to install mods great you need to download the 1.18.1 forge installer from forge's website, and then in your downloads double left click the file. if this does not work then go to the command line and: FOR WINDOWS right-click the file, go to Properties, navigate to the Security tab, and then click Edit and change the permissions so you have read write and execute permissons of the file, FOR MAC AND LINUX USERS use the chmod command and it looks like this chmod 700 filename. then go to minecraft and launch forge. When you get to the screen that lets you click miltiplayer or singleplayer find the tab on that screen that says mods and click it. then click open mods folder and find the .jar file and move it to the mods folder and then click done and click done click singleplayer and create new creative world and then launch the game and then you have the mod waiting for you in the misc. tab of the creative menu.